Refactor MNC viewer modules
